[MOV] moved '_message_post_helper' method in website_mail to payment for using sale module and working that functionality on the quote view without website
